Remove `pymc` based model
Removed the `pymc` based implementation of `vaxflux`'s core model to
only favor the `numpyro` based implementation. Also greatly simplified
the exports by removing dedicated modules which makes the package
unnecessarily complex. Users can now directly reference the core
elements of the package as `vaxflux.*`.

- Removed `UptakeModel` and associated PyMC based implementations of the
  `vaxflux` model.
- Removed all dedicated modules and consolidated the exports to
  `vaxflux.*`, with the exception of `vaxflux.data`.
- Removed older demo notebooks and updated the still relevant ontes to
  account for the export adjustments.
- Removed the API documentation generation since API reference is now
  simple.

Closes #172.
