+++ .gitignore
@@ -5,7 +5,6 @@ demos/**/*.png
 demos/scripts/
 
 # Generated documentation elements
-docs/api/
 docs/changelog.md
 docs/contributing.md
 docs/index.md
