+++ docs/api/data.md
@@ -1,0 +1,3 @@
+# Data
+
+::: vaxflux.data
