+++ docs/api/vaxflux.md
@@ -1,0 +1,3 @@
+# Vaxflux
+
+::: vaxflux
