+++ mkdocs.yml
@@ -69,19 +69,10 @@ nav:
   - Getting Started: getting-started.md
   - Model Details: model-details.md
   - Demos:
-      - IDD Demo: demos/idd-demo.ipynb
-      - Synthetic Data Daily: demos/synthetic-data-daily.ipynb
-      - Synthetic Data: demos/synthetic-data.ipynb
-      - Uptake Model: demos/uptake-model.ipynb
       - Vaxflux Model: demos/vaxflux-model.ipynb
       - Season Only Vaxflux Model: demos/season-only-vaxflux-model.ipynb
   - API Reference:
       - Vaxflux: api/vaxflux.md
-      - Covariates: api/covariates.md
-      - Curves: api/curves.md
       - Data: api/data.md
-      - Dates: api/dates.md
-      - Interventions: api/interventions.md
-      - Uptake: api/uptake.md
   - Contributing: contributing.md
   - Changelog: changelog.md
