+++ pyproject.toml
@@ -7,7 +7,6 @@ license = { text = "MIT" }
 requires-python = ">=3.11,<3.13"
 dependencies = [
     "numpy>=1.17.0,<2",
-    "pymc>5.16.2",
     "scipy>=1.14.1",
     "arviz>=0.19.0",
     "requests>=2.32.3",
@@ -111,10 +110,6 @@ max-args = 8  # Default is 5
 [tool.mypy]
 plugins = ["numpy.typing.mypy_plugin"]
 exclude = "demos/*"
-
-[[tool.mypy.overrides]]
-module = ["pymc.*"]
-ignore_missing_imports = true
 
 [[tool.mypy.overrides]]
 module = ["numpyro.*"]
